Issue #985 Move suggest and update button into more appropriate services detail component
